+++ src/app/components/providers/index.tsx
@@ -1,0 +1,4 @@
+import SessionProvider from './session'
+import {DynamicProvider} from './dynamic'
+
+export {SessionProvider, DynamicProvider}
